[ref]: separate vacancy service into DML, search services
