+++ vacancy/.env.example
@@ -11,4 +11,5 @@ POSTGRES_DB="vacancy"
 REDIS_HOST="localhost"
 REDIS_PORT=6379
 REDIS_DB=0
-REDIS_CACHE_TTL=5
+REDIS_CACHE_TTL=PT5M
+REDIS_TIMEOUT=0.3
