+++ vacancy/src/core/config/cache.py
@@ -1,3 +1,5 @@
+from datetime import timedelta
+
 from pydantic import Field
 
 from src.core.config.base import BaseConfig
@@ -7,4 +9,6 @@ class RedisConfig(BaseConfig):
     redis_host: str = Field(default="localhost", validation_alias="REDIS_HOST")
     redis_port: int = Field(default=6379, validation_alias="REDIS_PORT")
     redis_db: int = Field(default=0, validation_alias="REDIS_DB")
-    redis_ttl: int = Field(default=5, validation_alias="REDIS_CACHE_TTL")
+
+    redis_ttl: timedelta = Field(default=timedelta(minutes=5), validation_alias="REDIS_CACHE_TTL")
+    redis_timeout: float = Field(default=0.3, validation_alias="REDIS_TIMEOUT")
